+++ src/main/java/lagatrix/server/connection/requester/RequestOS.java
@@ -5,6 +5,7 @@ import lagatrix.server.entities.actions.ActionsEnum;
 import lagatrix.server.entities.connection.Request;
 import lagatrix.server.entities.connection.Response;
 import lagatrix.server.exceptions.LagatrixException;
+import lagatrix.server.file.log.LogContoller;
 import lagatrix.server.manager.OSManager;
 import lagatrix.server.tools.command.CommandExecutor;
 
@@ -18,8 +19,8 @@ public class RequestOS extends RequestManager {
     
     private OSManager manager;
 
-    public RequestOS(AESCommunicator communicator, CommandExecutor executor) {
-        super(communicator, executor);
+    public RequestOS(AESCommunicator communicator, CommandExecutor executor, LogContoller logger) {
+        super(communicator, executor, logger);
         this.manager = new OSManager(executor);
     }
 
@@ -29,6 +30,7 @@ public class RequestOS extends RequestManager {
         
         if (request.getAction() == ActionsEnum.GET){
             response.setResponse(manager.obtainOSInformation());
+            logger.info(communicator.getClientIp(), request.getAction(), "operative system");
         }
         
         communicator.sendResponse(response);