+++ src/main/java/lagatrix/server/connection/requester/RequestProcess.java
@@ -5,6 +5,7 @@ import lagatrix.server.entities.actions.ActionsEnum;
 import lagatrix.server.entities.connection.Request;
 import lagatrix.server.entities.connection.Response;
 import lagatrix.server.exceptions.LagatrixException;
+import lagatrix.server.file.log.LogContoller;
 import lagatrix.server.manager.ProcessManager;
 import lagatrix.server.tools.command.CommandExecutor;
 
@@ -18,23 +19,26 @@ public class RequestProcess extends RequestManager {
     
     private ProcessManager manager;
 
-    public RequestProcess(AESCommunicator communicator, CommandExecutor executor) {
-        super(communicator, executor);
+    public RequestProcess(AESCommunicator communicator, CommandExecutor executor, LogContoller logger) {
+        super(communicator, executor, logger);
         this.manager = new ProcessManager(executor);
     }
 
     @Override
     public void determineRequest(Request request) throws LagatrixException {
         Response response = new Response();
+        String message = null;
         
         // Determine the request of process.
         if (request.getAction() == ActionsEnum.GET){
             response.setResponse(manager.getProcess());
         } else if (request.getAction() == ActionsEnum.DELETE) {
             manager.killProcess((int) request.getParams()[0]);
+            message = String.format("the PID of process who kill is %d", request.getParams()[0]);
             response.setCorrectResult(true);
         }
         
+        logger.info(communicator.getClientIp(), request.getAction(), "process", message);
         communicator.sendResponse(response);
     }
 }