+++ CHANGELOG.md
@@ -35,6 +35,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 - LASRV010 - 04/04/2023 - Changes in the manage of response in the requests class.
 - LASRV005 - 04/04/2023 - Changes in the OSInformation, it have the package manager components.
 - LASRV010 - 05/04/2023 - Refractor in the communicators exceptions.
+- LASRV012 - 11/04/2023 - Implements LogController class in the program.
 
 ### Fixed